到昇腾NPU的原生迁移支持

+++ ultralytics/models/yolo/detect/val.py
@@ -264,7 +264,7 @@ class DetectionValidator(BaseValidator):
             self.jdict = []
             self.metrics.clear_stats()
         if self.args.plots and RANK > -1:
-            matrix = torch.as_tensor(self.confusion_matrix.matrix, device=self.device)
+            matrix = torch.as_tensor(self.confusion_matrix.matrix, dtype=torch.float32, device=self.device)
             dist.reduce(matrix, dst=0, op=dist.ReduceOp.SUM)
             if RANK == 0:
                 self.confusion_matrix.matrix = matrix.cpu().numpy()
